requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
dotenv==0.9.9
filelock==3.18.0
fsspec==2025.3.2
idna==3.10
packaging==25.0
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.3
tk==0.1.0
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.4.0
transformers>=4.30.0
python-docx>=0.8.11
huggingface-hub==0.30.2
Django==5.2